Refactor MainActivity and DetailFragment +MainFragment implements a Factory Method to create fragment instances +Use a PagerAdapter in MainActivity to host the fragments +Use a BottomNavigationView in MainActivity to switch between fragments +New icons for the above +Return of the DetailActivity to host the DetailFragment
-The movie storage database needs to be reworked to support this architecture properly.
Specifically we need to store popular and top rated movies concurrently, and access
the correct database from each fragment. Currently both the fragments in MainActivity
will end up displaying the same list until we refresh from one of the fragments.
Then the other fragment will still display the older list, until it is refreshed as well,
when it is recreated by screen rotation or app exit and relaunch.
-Disable shared element transitions until the next commit